making progress on the rewrite

+++ base/src/ui/draw/types.hpp
@@ -8,6 +8,6 @@
 #include <cstdint>
 
 namespace gta_base::ui::draw {
-  using sub_id_t = std::int64_t;
+  using sub_id_t = std::size_t;
 }
 #endif //GTA_BASE_TYPES_29AE7D84D73F47D8A73B1E9D0CC7C1A2_HPP
